--- FlyOrDie/MenuScene.swift
@@ -1,14 +1,0 @@
-//
-//  MenuScene.swift
-//  FlyOrDie
-//
-//  Created by KBryan on 2016-02-15.
-//  Copyright © 2016 KBryan. All rights reserved.
-//
-
-import UIKit
-import SpriteKit
-
-class MenuScene: SKScene {
-
-}
